move harvest specific approach to harvest component
